requirements.txt
Flask
Flask-WTF
WTForms
opencv-python
numpy
ultralytics
shapely
openpyxl
pytube
yt-dlp
gunicorn
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
flask-ngrok
pyngrok
requests
python-gnupg
python-dotenv
waitress
dill
lap